server/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
opencv-python
numpy
ultralytics
requests
